* develop:
  update js
  update code home help update code index help fix more bugs
  update code help update code helpers update code settings update code search create DTO Fix more bugs
  search is tested fix upload file
  update code settings replace websocket flush cache update DTO update database
  update code catalogs final more bugs fixed
  update code catalogs update database more bugs fixed
  database fix cabinet final code more other bugs fix
  Create resources, fix more bugs

+++ app/Catalogs/Actions/CabinetAction.php
@@ -1,30 +1,40 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Catalogs\Actions;
 
 use App\Base\Actions\Action;
+use App\Catalogs\DTO\CabinetDTO;
+use App\Core\Contracts\ICatalog;
+use App\Core\Contracts\ICatalogExtented;
 use App\Models\Cabinet as Model;
 use App\Models\User;
+use Carbon\Carbon;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Cache;
 
-class CabinetAction extends Action
+final class CabinetAction extends Action implements ICatalog, ICatalogExtented
 {
     /**
-     * [result cabinet]
-     */
-    private array $response;
-
-    /**
      * [count user for cabinet]
+     *
+     * @var countUser
      */
     private int $countUser;
 
     /**
-     * [all cabinet with count items on page]
+     * [all cabinet cache with count items on page]
+     *
+     * @return array{data: Illuminate\Pagination\LengthAwarePaginator}
      */
     public function getAllPagesPaginate(): array
     {
-        $this->items = Model::orderBy('description', 'ASC')->paginate($this->page);
+        $this->currentPage = request()->get('page', 1);
+        $this->items = Cache::tags('cabinet')->remember('cabinet.'.$this->currentPage, Carbon::now()->addDay(), function () {
+            return Model::query()->orderBy('description', 'ASC')->paginate($this->page);
+        });
+
         $this->response =
         [
             'data' => $this->items,
@@ -36,22 +46,40 @@ class CabinetAction extends Action
     /**
      * [show one cabinet]
      */
-    public function show(int $id): Model
+    public function show(int $id): array
     {
-        $this->item = Model::findOrFail($id);
+        $this->item = Model::query()->find($id);
+        if (! $this->item) {
 
-        return $this->item;
+            return abort(404);
+        }
+        $this->response =
+        [
+            'data' => $this->item,
+        ];
+
+        return $this->response;
     }
 
     /**
      * [add new cabinet]
+     *
+     * @param  array  $request {description: string}
      */
     public function store(array $request): JsonResponse
     {
-        $this->item = Model::create($request);
+        $this->dataObject = new CabinetDTO(
+            $request['description']
+        );
+        $this->item = new Model();
+        $this->item->description = $this->dataObject->description;
+        $this->item->save();
         $this->response = [
-            'message' => 'Кабинет успешно добавлен!',
+            'message' => 'Кабинет успешно добавлен в очередь на размещение!',
+            'reload' => true,
         ];
+
+        Cache::tags('cabinet')->flush();
 
         return response()->success($this->response);
 
@@ -59,14 +87,30 @@ class CabinetAction extends Action
 
     /**
      * [update cabinet]
+     *
+     * @param  array  $request {description: string}
      */
     public function update(array $request, int $id): JsonResponse
     {
-        $this->item = Model::findOrFail($id);
-        $this->item->update($request);
+        $this->dataObject = new CabinetDTO(
+            $request['description']
+        );
+        $this->item = Model::query()->find($id);
+
+        if (! $this->item) {
+            $this->response = [
+                'message' => 'Кабинет не найден!',
+            ];
+
+            return response()->error($this->response);
+        }
+        $this->item->description = $this->dataObject->description;
+        $this->item->save();
         $this->response = [
-            'message' => 'Кабинет успешно обновлён!',
+            'message' => 'Кабинет успешно добавлен в очередь на обновление!',
         ];
+
+        Cache::tags('cabinet')->flush();
 
         return response()->success($this->response);
     }
@@ -74,9 +118,10 @@ class CabinetAction extends Action
     /**
      * [delete cabinet if there are no employees]
      */
-    public function delete(int $id): JsonResponse
+    public function destroy(int $id): JsonResponse
     {
-        $this->countUser = User::where('cabinet_id', $id)->count();
+        $this->countUser = User::query()->where('cabinet_id', $id)->count();
+
         if ($this->countUser > 0) {
             $this->response = [
                 'message' => 'Кабинет не может быть удален, так как у кабинета есть сотрудники!',
@@ -84,12 +129,22 @@ class CabinetAction extends Action
 
             return response()->error($this->response);
         }
-        $this->item = Model::findOrFail($id);
+        $this->item = Model::query()->find($id);
+
+        if (! $this->item) {
+            $this->response = [
+                'message' => 'Кабинет не найден!',
+            ];
+
+            return response()->error($this->response);
+        }
         $this->item->forceDelete();
         $this->response = [
-            'message' => 'Кабинет успешно удалён!',
+            'message' => 'Кабинет успешно поставлен в очередь на удаление!',
             'reload' => true,
         ];
+
+        Cache::tags('cabinet')->flush();
 
         return response()->success($this->response);
     }
