+++ app/Catalogs/Actions/CategoryAction.php
@@ -1,30 +1,41 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Catalogs\Actions;
 
 use App\Base\Actions\Action;
+use App\Base\Helpers\StringHelper;
+use App\Catalogs\DTO\CategoryDTO;
+use App\Core\Contracts\ICatalog;
+use App\Core\Contracts\ICatalogExtented;
 use App\Models\Category as Model;
 use App\Models\Help;
+use Carbon\Carbon;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Cache;
 
-class CategoryAction extends Action
+final class CategoryAction extends Action implements ICatalog, ICatalogExtented
 {
     /**
-     * [result category]
-     */
-    private array $response;
-
-    /**
-     * [count help for category]
+     * [count user for category]
+     *
+     * @var countUser
      */
     private int $countHelp;
 
     /**
-     * [all category with count items on page]
+     * [all category cache with count items on page]
+     *
+     * @return array{data: Illuminate\Pagination\LengthAwarePaginator}
      */
     public function getAllPagesPaginate(): array
     {
-        $this->items = Model::orderBy('description', 'ASC')->paginate($this->page);
+        $this->currentPage = request()->get('page', 1);
+        $this->items = Cache::tags('category')->remember('category.'.$this->currentPage, Carbon::now()->addDay(), function () {
+            return Model::query()->orderBy('description', 'ASC')->paginate($this->page);
+        });
+
         $this->response =
         [
             'data' => $this->items,
@@ -36,36 +47,71 @@ class CategoryAction extends Action
     /**
      * [show one category]
      */
-    public function show(int $id): Model
+    public function show(int $id): array
     {
-        $this->item = Model::findOrFail($id);
+        $this->item = Model::query()->find($id);
+        if (! $this->item) {
 
-        return $this->item;
+            return abort(404);
+        }
+        $this->response =
+        [
+            'data' => $this->item,
+        ];
+
+        return $this->response;
     }
 
     /**
      * [add new category]
+     *
+     * @param  array  $request {description: string}
      */
     public function store(array $request): JsonResponse
     {
-        Model::create($request);
+        $this->dataObject = new CategoryDTO(
+            $request['description']
+        );
+        $this->item = new Model();
+        $this->item->description = StringHelper::run($this->dataObject->description);
+        $this->item->save();
         $this->response = [
-            'message' => 'Категория успешно добавлена!',
+            'message' => 'Категория успешно добавлена в очередь на размещение!',
+            'reload' => true,
         ];
 
+        Cache::tags('category')->flush();
+
         return response()->success($this->response);
+
     }
 
     /**
      * [update category]
+     *
+     * @param  array  $request {description: string}
      */
     public function update(array $request, int $id): JsonResponse
     {
-        $this->item = Model::findOrFail($id);
-        $this->item->update($request);
+        $this->dataObject = new CategoryDTO(
+            $request['description']
+        );
+        $this->item = Model::query()->find($id);
+
+        if (! $this->item) {
+            $this->response = [
+                'message' => 'Категория не найдена!',
+            ];
+
+            return response()->error($this->response);
+        }
+        $this->item->description = StringHelper::run($this->dataObject->description);
+        $this->item->save();
         $this->response = [
-            'message' => 'Категория успешно обновлена!',
+            'message' => 'Категория успешно добавлена в очередь на обновление!',
         ];
+
+        Cache::tags('category')->flush();
 
         return response()->success($this->response);
     }
@@ -73,22 +119,32 @@ class CategoryAction extends Action
     /**
      * [delete category if there are no help]
      */
-    public function delete(int $id): JsonResponse
+    public function destroy(int $id): JsonResponse
     {
         $this->countHelp = Help::where('category_id', $id)->count();
         if ($this->countHelp > 0) {
             $this->response = [
                 'message' => 'Категория не может быть удалена, так как не удалены все заявки связанные с ней!',
-                'reload' => true,
             ];
 
             return response()->error($this->response);
         }
-        $this->item = Model::findOrFail($id);
+        $this->item = Model::query()->find($id);
+
+        if (! $this->item) {
+            $this->response = [
+                'message' => 'Категория не найдена!',
+            ];
+
+            return response()->error($this->response);
+        }
         $this->item->forceDelete();
         $this->response = [
-            'message' => 'Категория успешно удалена!',
+            'message' => 'Категория успешно поставлена в очередь на удаление!',
+            'reload' => true,
         ];
+
+        Cache::tags('category')->flush();
 
         return response()->success($this->response);
     }
